Start problem 4
Commence problem 4 by defining a min-simple helper function that returns the minimum number in a simple list containing letters and/or numbers. The function logic so far returns the only element in a list if the element is a number, otherwise, it returns a really large number. This works on the assumption that no list will contain a value larger than 10000000000000000000.