Complete helper function for problem 4
The remaining logic for the helper function has been added. I check if the first number in a list is less than the result of recursively calling my function on the rest of the list. If it is so, that number is returned, otherwise whatever new minimum is discovered during the numerous recursive calls is returned.