Updated helper function for simple minimum
I scrapped the previous implementation of the minimum finder function in favor of a more elegant solution that should integrate better with the rest of the functionality. The logic here is to keep track of the current minimum value at all times, initializing it to false at the beginning. If the first element in the list is a number and the current minimum value is still the default false, recursively call the function on the rest of the list with the current minimum replaced as the first element in the list.