Finished new helper function implementation
The rest of the function logic checks if the first element in the list is less than the current minimum (already overwritten with a numerical value), it recursively calls the function if this is the case, using this new value as the minimum. Otherwise, the function is called with the previous minimum value on the rest of the list.