Start new helper function
I defined a min-above-num helper function that returns the minimum number in a list that is above whatever number is passed into it. The logic defaults the current minimum to false and returns that if the list is empty. If the first element in the list is not a number, the function is called recursively on the rest of the list with the current minimum default value of false.