Continue second helper function
In the event that the first element in the list is a number, I check to confirm that the minimum value I'm comparing to is also a number. If it is not a number, I also check whether the current minimum has been assigned a value other than the default (false). If this is not the case, I make a recursive call using different parts of the list and the other minimum value that's being compared; if it is the case, the function is called recursively with the following parameters: the rest of the list, the smaller value between the first element in the list and the current minimum value, the other minimum value that the list is being compared to.