Complete second helper function
If the first element in the list is less than or equal to the number that's being compared to, call the function recursively on the rest of the list with similar parameters, otherwise, call the function with updated parameters if the current minimum is not a number. If the previous condition is not met, call the function once more on the rest of the list as first parameter, the smaller element between the first element in the list and current minimum and the number in consideration as the last parameter.